Pulled implementations for search empty cell and saving a cell back into the more general Excel class
